Add PrepTime endpoint for adding preptime for use when assign courses to faculty

+++ WhoteachswhatJPA/src/main/java/com/seneca/repository/PrepTimeDao.java
@@ -8,15 +8,15 @@ public interface PrepTimeDao
 {
    public List<PrepTime> getAll();
 
-   public PrepTime getById(Long id);
+   public PrepTime getById(Integer id);
 
    public void update(PrepTime entity);
 
    public void create(PrepTime entity);
 
-   public void delete(Long id);
+   public void delete(Integer id);
 
-   public List<PrepTime> search(PrepTime search, Long first, Integer maxItems);
+   public List<PrepTime> search(PrepTime search, Integer first, Integer maxItems);
 
    public Long getCount(PrepTime search);
 }
